requirements.txt
autopep8==1.4.4
boto3==1.9.247
botocore==1.12.247
certifi==2019.9.11
chardet==3.0.4
docutils==0.15.2
idna==2.8
jmespath==0.9.4
pika==1.1.0
Pillow==6.2.0
pycodestyle==2.5.0
python-dateutil==2.8.0
python-image-resize==1.1.2
python-resize-image==1.1.19
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.6
